[ARC-595] Move commands to sub package
branch: feature/ARC-595-mpyl-basic-cli
