+++ src/mpyl/cli/__init__.py
@@ -1,1 +1,1 @@
-"""CLI commands"""
+"""Command Line Interface for MPyL"""
